Refactor echo estimation modules and remove legacy correlation imports
